add the exception of Tag API

+++ src/controllers/tag.ts
@@ -1,5 +1,8 @@
+import mongoose from 'mongoose';
 import { Request, Response, NextFunction } from 'express';
 import TagModel, { ITag } from '../models/Tag';
+import TagNotFoundException from '../exceptions/TagNotFoundException';
+import IdIsNotValidException from '../exceptions/IdIsNotValidException';
 
 export const getTags = async (
   req: Request,
@@ -19,32 +22,28 @@ export const getTagById = async (
   res: Response,
   next: NextFunction
 ) => {
-  try {
-    const {id} = req.params;
-    const tag = await TagModel.findById(id);
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    next(new IdIsNotValidException(id));
+  }
+
+  const tag = await TagModel.findById(id);
+  if (tag) {
     res.send(tag);
-  } catch (e) {
-    next(e);
+  } else {
+    next(new TagNotFoundException(id));
   }
 };
 
-export const createTag = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const tag: ITag = req.body;
-    const isExist = await TagModel.exists({ name: tag.name });
-    if (isExist) {
-      res.send('the tag is is exists.');
-    } else {
-      const newTag = new TagModel(tag);
-      const saveTag = await newTag.save();
-      res.send(saveTag);
-    }
-  } catch (e) {
-    next(e);
+export const createTag = async (req: Request, res: Response) => {
+  const tag: ITag = req.body;
+  const isExist = await TagModel.exists({ name: tag.name });
+  if (isExist) {
+    res.send('the tag is is exists.');
+  } else {
+    const newTag = new TagModel(tag);
+    const saveTag = await newTag.save();
+    res.send(saveTag);
   }
 };
 
@@ -53,17 +52,17 @@ export const updateTagById = async (
   res: Response,
   next: NextFunction
 ) => {
-  try {
-    const {id} = req.params;
-    const newTag: ITag = req.body;
-    const updateTag = await TagModel.findByIdAndUpdate(id, newTag);
-    if (updateTag) {
-      res.send(`update success`);
-    } else {
-      res.send(`update failed`);
-    }
-  } catch (e) {
-    next(e);
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    next(new IdIsNotValidException(id));
+  }
+
+  const newTag: ITag = req.body;
+  const updateTag = await TagModel.findByIdAndUpdate(id, newTag);
+  if (updateTag) {
+    res.send(updateTag);
+  } else {
+    next(new TagNotFoundException(id));
   }
 };
 
@@ -72,15 +71,15 @@ export const deleteTagById = async (
   res: Response,
   next: NextFunction
 ) => {
-  try {
-    const {id} = req.params;
-    const tag = await TagModel.findByIdAndDelete(id);
-    if (tag) {
-      res.send(`delete success`);
-    } else {
-      res.send(`delete failed`);
-    }
-  } catch (e) {
-    next(e);
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    next(new IdIsNotValidException(id));
+  }
+
+  const tag = await TagModel.findByIdAndDelete(id);
+  if (tag) {
+    res.send(tag);
+  } else {
+    next(new TagNotFoundException(id));
   }
 };
